requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.7
mysql-connector-python==8.0.32
Flask-Migrate==4.0.4
PyJWT==2.6.0
bcrypt==4.0.1
Werkzeug==2.2.3